fix test cases

+++ tendercuts/app/inventory/tests/unit/test_inventory_controller.py
@@ -17,9 +17,9 @@ def mock_user():
 
 @pytest.fixture()
 def mock_inv():
-    try:
-        entry = Graminventory.objects.get(
-            store_id=1, product_id=193, date=datetime.date.today())
+    entry = Graminventory.objects.filter(
+        store_id=1, product_id=193, date=datetime.date.today())
+    if entry:
         entry.delete()
     except:
         pass
@@ -52,7 +52,8 @@ def test_today_qty_upload(mock_inv, mock_user):
 
     log = InventoryController(mock_inv).process_inventory_request(req)
 
-    new_inv = Graminventory.objects.get(pk=mock_inv.graminventoryid)  # type: Graminventory
+    new_inv = Graminventory.objects.get(
+        pk=mock_inv.graminventoryid)  # type: Graminventory
     assert new_inv.qty == 0
 
     assert log.sku == req.sku
@@ -76,7 +77,8 @@ def test_tomo_qty_upload(mock_inv, mock_user):
 
     log = InventoryController(mock_inv).process_inventory_request(req)
 
-    new_inv = Graminventory.objects.get(pk=mock_inv.graminventoryid)  # type: Graminventory
+    new_inv = Graminventory.objects.get(
+        pk=mock_inv.graminventoryid)  # type: Graminventory
     assert new_inv.forecastqty == 0
 
     assert log.sku == req.sku
